requirements.txt
requests>=2.31.0
python-dateutil>=2.8.2
streamlit>=1.32.0
plotly>=5.20.0
pandas>=2.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
python-jose[cryptography]>=3.3.0
eth-account>=0.11.0
slowapi>=0.1.9
python-multipart>=0.0.9
httpx>=0.27.0
